orchestration-agent-test/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.0
pydantic==2.9.2
